+++ presets/bigCircles/src/index.ts
@@ -7,13 +7,13 @@ import { loadOpacityUpdater } from "tsparticles-updater-opacity";
 import { loadOutModesUpdater } from "tsparticles-updater-out-modes";
 import { loadSizeUpdater } from "tsparticles-updater-size";
 
-export function loadBigCirclesPreset(tsParticles: Main): void {
-    loadCircleShape(tsParticles);
-    loadColorUpdater(tsParticles);
-    loadSizeUpdater(tsParticles);
-    loadOpacityUpdater(tsParticles);
-    loadOutModesUpdater(tsParticles);
-    loadEmittersPlugin(tsParticles);
+export async function loadBigCirclesPreset(tsParticles: Main): Promise<void> {
+    await loadCircleShape(tsParticles);
+    await loadColorUpdater(tsParticles);
+    await loadSizeUpdater(tsParticles);
+    await loadOpacityUpdater(tsParticles);
+    await loadOutModesUpdater(tsParticles);
+    await loadEmittersPlugin(tsParticles);
 
     tsParticles.addPreset("bigCircles", options);
     tsParticles.addPreset("big-circles", options);