+++ components/jquery/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.3](https://github.com/matteobruni/tsparticles/compare/jquery-particles@2.0.0-beta.2...jquery-particles@2.0.0-beta.3) (2021-12-07)
+
+**Note:** Version bump only for package jquery-particles
+
+
+
+
+
 # [2.0.0-beta.2](https://github.com/matteobruni/tsparticles/compare/jquery-particles@1.37.5...jquery-particles@2.0.0-beta.2) (2021-12-04)