+++ updaters/size/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.4](https://github.com/matteobruni/tsparticles/compare/tsparticles-updater-size@2.0.0-beta.3...tsparticles-updater-size@2.0.0-beta.4) (2021-12-07)
+
+**Note:** Version bump only for package tsparticles-updater-size
+
+
+
+
+
 # [2.0.0-beta.3](https://github.com/matteobruni/tsparticles/compare/tsparticles-updater-size@2.0.0-beta.2...tsparticles-updater-size@2.0.0-beta.3) (2021-12-04)
 
 **Note:** Version bump only for package tsparticles-updater-size