+++ utils/configs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.2](https://github.com/matteobruni/tsparticles/compare/tsparticles-demo-configs@1.0.0-beta.1...tsparticles-demo-configs@1.0.0-beta.2) (2021-12-07)
+
+**Note:** Version bump only for package tsparticles-demo-configs
+
+
+
+
+
 # 1.0.0-beta.1 (2021-12-04)