+++ utils/webpack/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.2](https://github.com/matteobruni/tsparticles/compare/webpack-tsparticles-plugin@1.0.0-beta.1...webpack-tsparticles-plugin@1.0.0-beta.2) (2021-12-07)
+
+**Note:** Version bump only for package webpack-tsparticles-plugin
+
+
+
+
+
 # [1.0.0-beta.1](https://github.com/matteobruni/tsparticles/compare/webpack-tsparticles-plugin@1.0.0-beta.0...webpack-tsparticles-plugin@1.0.0-beta.1) (2021-12-04)